requirements.txt
aiohttp==3.9.1
python-dotenv==1.0.0
rich==13.7.0
pydantic==2.5.2
pydantic-settings==2.1.0
tenacity==8.2.3
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0